Change PurePursue command package
